+++ sources/logdog/Cargo.toml
@@ -8,7 +8,7 @@ publish = false
 
 [dependencies]
 flate2 = "1.0"
-snafu = "0.6"
+snafu = { version = "0.6", features = ["backtraces-impl-backtrace-crate"] }
 tar = { version = "0.4", default-features = false }
 tempfile = { version = "3.1.0", default-features = false }
